// src/GitHub.Interface.ts
export interface GitHubUser {
    login: string
    id: number
    avatar_url: string
    html_url: string
    name: string | null
    company: string | null
    blog: string | null
    location: string | null
    email: string | null
    bio: string | null
    twitter_username: string | null
    public_repos: number
    public_gists: number
    followers: number
    following: number
    created_at: string
    updated_at: string
  }
  
  export interface Repository {
    id: number
    name: string
    full_name: string
    html_url: string
    description: string | null
    fork: boolean
    created_at: string
    updated_at: string
    pushed_at: string
    homepage: string | null
    size: number
    stargazers_count: number
    watchers_count: number
    language: string | null
    forks_count: number
    open_issues_count: number
    license: {
      key: string
      name: string
      url: string
    } | null
    topics: string[]
    visibility: string
    default_branch: string
  }
  
  export interface CommitData {
    date: string
    commits: number
  }
  
  